småting
la til en kapp for å kunne ta opp og fjerne info punktene. tok også å delte koden opp i flere deler
