+++ Index.html
@@ -1,4 +1,3 @@
-
 <!DOCTYPE html>
 <html lang="no">
 <head>
@@ -17,9 +16,9 @@
 </head>
 <body>
     <div id="map"></div>
+    <button id="showRouteInfo">Vis Ruteinformasjon</button>
 
     <!-- Link to external JavaScript file -->
     <script type="module" src="script.js"></script>
-
 </body>
 </html>
